add user id to donation form
